Move struct 'TestCase' into new header 'test_case.hpp'
